build: make TestedTrek into a Maven module and update tests to JUnit 5
